set a queue status to allow queue to be started or stopped.
