+++ internal/botaccess/admincommands.go
@@ -7,6 +7,7 @@ import (
 
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 	"github.com/josh1248/nusc-queue-bot/internal/dbaccess"
+	"github.com/josh1248/nusc-queue-bot/internal/queuestatus"
 	"github.com/josh1248/nusc-queue-bot/internal/types"
 )
 
@@ -29,6 +30,16 @@ var AdminCommands = []types.AcceptedCommands{
 		Command:     "done",
 		Description: "remove the first person from the queue once they have finished their photo-taking.",
 		Handler:     RemoveFirstInQueueCommand,
+	},
+	{
+		Command:     "stopqueue",
+		Description: "Stop admitting perople into the queue.",
+		Handler:     StopQueueCommand,
+	},
+	{
+		Command:     "startqueue",
+		Description: "Start admitting perople into the queue.",
+		Handler:     StartQueueCommand,
 	},
 	{
 		Command:     "adminlist",
@@ -55,6 +66,108 @@ var AdminCommands = []types.AcceptedCommands{
 		Description: "Explains the main functionalities of the bot.",
 		Handler:     AdminHelpCommand,
 	},
+}
+
+func SeeQueueCommand(userMessage tgbotapi.Update, bot *tgbotapi.BotAPI) (feedback string) {
+	queueUsers, err := dbaccess.CheckQueueContents()
+	if err != nil {
+		feedback = "Something went wrong when accessing the queue... blame @joshtwo."
+		log.Println(err)
+	}
+
+	userToStr := func(user types.QueueUser) string {
+		return fmt.Sprintf("@%s %s\n", user.UserHandle, user.Joined_at.Format("15:04:05"))
+	}
+
+	var sb strings.Builder
+	sb.WriteString(fmt.Sprintf("Total queue length: %v \n\nName		Joined at\n ", len(queueUsers)))
+	for _, user := range queueUsers {
+		sb.WriteString(userToStr(user))
+	}
+
+	feedback = sb.String()
+	return feedback
+}
+
+func StopQueueCommand(userMessage tgbotapi.Update, bot *tgbotapi.BotAPI) (feedback string) {
+	queuestatus.SetQueueClose()
+	return "queue successfully stopped."
+}
+
+func StartQueueCommand(userMessage tgbotapi.Update, bot *tgbotapi.BotAPI) (feedback string) {
+	queuestatus.SetQueueOpen()
+	return "queue successfully opened."
+}
+
+// To update: should be variable based on whether you have joined the queue.
+func HowLongCommand(userMessage tgbotapi.Update, bot *tgbotapi.BotAPI) (feedback string) {
+	isInQueue, queueLength, err := dbaccess.CheckQueueLength(userMessage.SentFrom().UserName)
+	if err != nil {
+		feedback = "Something went wrong when accessing the queue... blame @joshtwo."
+		log.Println(err)
+		return feedback
+	}
+
+	info := func(queueLength int) string {
+		if queueLength == 1 {
+			return fmt.Sprintf("%s %d %s", "is", queueLength, "group")
+		} else {
+			return fmt.Sprintf("%s %d %s", "are", queueLength, "groups")
+		}
+	}
+
+	if isInQueue {
+		feedback = fmt.Sprintf("There %s in front of you.", info(queueLength-1))
+	} else {
+		feedback = fmt.Sprintf("There %s in the queue now. (Join the queue with /join.)", info(queueLength))
+	}
+	return feedback
+}
+
+func PingCommand(userMessage tgbotapi.Update, bot *tgbotapi.BotAPI) (feedback string) {
+	chatID, err := dbaccess.NotifyQueue(1)
+	if err != nil {
+		feedback = "You failed to kick the first person: " + err.Error()
+		log.Printf("Error sending message, %v\n", err)
+		return feedback
+	}
+
+	msg := tgbotapi.NewMessage(chatID, "Hey, you are the first person in queue! get moving :D")
+	_, err = bot.Send(msg)
+	if err != nil {
+		feedback = "You failed to kick the first person: " + err.Error()
+		log.Printf("Error sending message %v\n", err)
+		return feedback
+	}
+
+	feedback = "First person in queue notified."
+	return feedback
+}
+
+func KickCommand(userMessage tgbotapi.Update, bot *tgbotapi.BotAPI) (feedback string) {
+	if len(userMessage.Message.Text) < 7 {
+		feedback = "input the username to kick. Example: /kick @userABC"
+		return feedback
+	}
+
+	telegramHandle := userMessage.Message.Text[7:]
+	chatID, err := dbaccess.KickPerson(telegramHandle)
+	if err != nil {
+		feedback = "You failed to kick the first person: " + err.Error()
+		log.Printf("Error sending message %v\n", err)
+		return feedback
+	}
+
+	msg := tgbotapi.NewMessage(chatID, "You have been kicked from the queue.")
+	_, err = bot.Send(msg)
+	if err != nil {
+		feedback = "You failed to kick " + telegramHandle + " : " + err.Error()
+		log.Printf("Error sending message %v\n", err)
+		return feedback
+	}
+
+	feedback = "Successfully kicked " + telegramHandle
+	return feedback
 }
 
 func AdminHelpCommand(userMessage tgbotapi.Update, bot *tgbotapi.BotAPI) (feedback string) {
